added extra checks for admin

+++ app/Http/Controllers/postCrudController.php
@@ -16,19 +16,24 @@ class PostCrudController extends Controller
 
     public function update(Request $request){
         $post = Post::where('id', $request->currentId)->first();
-        $post->update(array('title' => $request->title));
-        $post->update(array('content' => $request->content));
-        if ($request->file('file')) {
-            $OldFile = $post->file;
-            $filePath = '/public/postsImg/' . $OldFile;
-            Storage::delete($filePath);   //delete the old file
-            $image = $request->file('file');
-            $imageName = time() . '.' . $image->getClientOriginalName();
-            $path = $request->file('file')->storeAs('postsImg', $imageName, 'public');//save the new file
-            $file = $imageName;
-            $post->update(array('file' => $file));
+        if (Auth::user()->id == $post->user->id || Auth::user()->hasRole('admin')){
+            $post->update(array('title' => $request->title));
+            $post->update(array('content' => $request->content));
+            if ($request->file('file')) {
+                $OldFile = $post->file;
+                $filePath = '/public/postsImg/' . $OldFile;
+                Storage::delete($filePath);   //delete the old file
+                $image = $request->file('file');
+                $imageName = time() . '.' . $image->getClientOriginalName();
+                $path = $request->file('file')->storeAs('postsImg', $imageName, 'public');//save the new file
+                $file = $imageName;
+                $post->update(array('file' => $file));
+            }
+            return redirect()->back()->withSuccess('Post succesfully edited!');
         }
-        return redirect()->back()->withSuccess('Post succesfully edited!');
+        else {
+            return view('layouts.noPermission');
+        }
     }
 
     public function delete($id){
@@ -48,15 +53,15 @@ class PostCrudController extends Controller
 
     public function store(Request $request) {
         //this gives us the currently logged in user
-        $user = $request->user();
-        $formdata = $request->all();
-        if ($request->file('file')) {
-            $image = $request->file('file');
-            $imageName = time() . '.' . $image->getClientOriginalName();
-            $path = $request->file('file')->storeAs('postsImg', $imageName, 'public');
-            $formdata['file'] = $imageName;
-        }
-        $user->posts()->create($formdata);
-        return redirect()->back()->withSuccess('Post succesfully created!');
+            $user = $request->user();//Auth==user() is also possible
+            $formdata = $request->all();
+            if ($request->file('file')) {
+                $image = $request->file('file');
+                $imageName = time() . '.' . $image->getClientOriginalName();
+                $path = $request->file('file')->storeAs('postsImg', $imageName, 'public');
+                $formdata['file'] = $imageName;
+            }
+            $user->posts()->create($formdata);
+            return redirect()->back()->withSuccess('Post succesfully created!');
     }
 }
